backend/requirements.txt
# FastAPI Stack
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.5.0
pydantic-settings>=2.1.0  # Add this for BaseSettings
python-multipart>=0.0.6
httpx>=0.27.0
itsdangerous>=2.1.2
python-jose[cryptography]>=3.3.0

# Database & Supabase
supabase>=2.3.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0  # PostgreSQL async driver
psycopg2-binary>=2.9.9  # PostgreSQL sync driver (for migrations)

# LiveKit
livekit-agents[openai,silero,turn-detector]~=1.0
livekit-api
livekit-plugins-google
livekit-plugins-noise-cancellation

# Previous Dependencies
duckduckgo-search
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dateutil
python-multipart